Move product registration stuff into ProductRegistry
