+++ api/src/org/labkey/api/data/Container.java
@@ -55,7 +55,6 @@ import org.labkey.api.security.permissions.Permission;
 import org.labkey.api.security.permissions.ReadPermission;
 import org.labkey.api.security.permissions.UpdatePermission;
 import org.labkey.api.security.roles.Role;
-import org.labkey.api.settings.AdminConsole;
 import org.labkey.api.settings.AppProps;
 import org.labkey.api.settings.LookAndFeelProperties;
 import org.labkey.api.settings.ProductFeature;
@@ -1044,7 +1043,7 @@ public class Container implements Serializable, Comparable<Container>, Securable
     public Boolean getAuditCommentsRequired()
     {
         Map<String, String> props = PropertyManager.getProperties(this, AUDIT_SETTINGS_PROPERTY_SET_NAME);
-        if (!AdminConsole.isProductFeatureEnabled(ProductFeature.DataChangeCommentRequirement))
+        if (!ProductRegistry.isProductFeatureEnabled(ProductFeature.DataChangeCommentRequirement))
             return false;
         return Boolean.parseBoolean(props.getOrDefault(REQUIRE_USER_COMMENTS_PROPERTY_NAME, "false"));
     }
@@ -1665,7 +1664,7 @@ public class Container implements Serializable, Comparable<Container>, Securable
         if (isWorkbook())
             return false;
 
-        return AdminConsole.isProductFeatureEnabled(feature);
+        return ProductRegistry.isProductFeatureEnabled(feature);
     }
 
     public boolean isFeatureEnabled(ProductFeature feature)
